short-form-factory/requirements.txt
google-generativeai
python-dotenv
requests
moviepy
google-auth-oauthlib
google-api-python-client
flask
flask-cors
